Perl install works without setting PERL5LIB

+++ pkg/platform/runtime/camel.go
@@ -189,6 +189,17 @@ func (cr *CamelRuntime) PostUnpackArtifact(artf *HeadChefArtifact, tmpRuntimeDir
 		}
 	}
 
+	tmpRelocFile := filepath.Join(tmpRuntimeDir, archiveName, "support/reloc.txt")
+	if fileutils.FileExists(tmpRelocFile) {
+		target := filepath.Join(installDir, "support/reloc.txt")
+		if fail := fileutils.MkdirUnlessExists(filepath.Dir(target)); fail != nil {
+			return fail
+		}
+		if err := os.Rename(tmpRelocFile, target); err != nil {
+			return FailRuntimeInstallation.Wrap(err)
+		}
+	}
+
 	if err := os.RemoveAll(tmpRuntimeDir); err != nil {
 		logging.Error("removing %s after unpacking runtime: %v", tmpRuntimeDir, err)
 		return FailRuntimeInstallation.New("installer_err_runtime_rm_installdir", tmpRuntimeDir)
@@ -214,14 +225,23 @@ func Relocate(metaData *MetaData, cb func()) *failures.Failure {
 	if len(prefix) == 0 || prefix == metaData.Path {
 		return nil
 	}
-
 	logging.Debug("relocating '%s' to '%s'", prefix, metaData.Path)
 	binariesSeparate := rt.GOOS == "linux" && metaData.RelocationTargetBinaries != ""
+
+	relocFilePath := filepath.Join(metaData.Path, "support", "reloc.txt")
+	relocMap := map[string]bool{}
+	if fileutils.FileExists(relocFilePath) {
+		relocMap = loadRelocationFile(relocFilePath)
+	}
 
 	// Replace plain text files
 	err := fileutils.ReplaceAllInDirectory(metaData.Path, prefix, metaData.Path,
 		// Check if we want to include this
 		func(p string, contents []byte) bool {
+			suffix := strings.TrimPrefix(p, metaData.Path)
+			if relocMap[suffix] {
+				return true
+			}
 			if !strings.HasSuffix(p, constants.RuntimeMetaFile) && (!binariesSeparate || !fileutils.IsBinary(contents)) {
 				cb()
 				return true
